# src/main/java/com/samuelbraga/pierpont/models/OperationType.java
package com.samuelbraga.pierpont.models;

import jakarta.persistence.*;
import lombok.*;

@Entity
@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
@Table(name = "operation_types")
public class OperationType {
  @Id
  @GeneratedValue(strategy = GenerationType.IDENTITY)
  private Long id;

  private String description;
}